requirements.txt
flask==3.0.2
google-generativeai==0.8.5
python-dotenv==1.0.1
gunicorn==21.2.0 